Fix #4 js client connection timeout exception
If the server closes the connection without a reply, eg. because a
timeout was hit before the MessageBus.long_polling_interval or due to a
server restart, the success callback is called with null data.
The code no handles this gracefully instead of raising a TypeError
inside jQuery.each trying to call object.length on messages.
This case is not counted as an error, because we usually can reconnect
just fine instantly.