Send flush message to JS when position is ahead
In certain circumstances, a JS client can get channel positions that are
ahead of the message bus. When this happens, the client will not recieve
any more messages until the page is refreshed. This patch fixes that by:

 - (middleware) delivering a message on the __flush channel when the
   condition is detected
 - (client) resetting all subscriptions to -1 when the __flush message
   is recieved

Note: The 'ahead' condition happens when the Redis DB is flushed.
Although this is not a common operation, flush-resiliency is a desirable
property of many applications using Redis; this change improves the
flush resiliency of MessageBus from "new messages will be lost for an
indeterminate period of time" to "all old messages are lost, plus those
that happen until the client next polls".