Simple attempt at bridging message delivery race
In the case that a message is published while we're calculating the backlog, we previously missed it. If the backlog turned out to be empty, we would have begun waiting for new messages, but the published message could have been delivered to the connection manager prior to informing it of the client. This would result in the message failing to be delivered until our long-poll timeout.

This adds an in-memory backlog per client which is filled while we calculate the backlog in Redis. The combined (deduplicated) backlog is then considered for delivery, closing the gap. Any messages received between collecting the message buffer and delivering the backlog will become part of the Redis backlog response on the next request.