Fix concurrency benchmark example
This changes the benchmark to use [wrk](https://github.com/wg/wrk)
instead of apache bench for testing concurrency. This is required
because we need to generate a unique client id or the server will
immediately recycle the client's previous connection, making the
benchmark useless. Also added a sample puma config.