Port the test suite to MiniTest
This actually uses minitest/spec.  Hopefully that's fine.

This replaces rspec with minitest in the Gemfile, and removes
the redis, rack, rbtrace, guard-rspec, and rb-notify gems from
the Gemfile.  redis and rack in the gemspec and do not need to
be specified twice, guard-rspec seems to be related to rspec,
and rbtrace and rb-inotify don't seem to be needed to run the
tests.

Most of this was done mechanically using the same script I
used for converting Sequel's and all of my other libraries'
specs from rspec to minitest.

I used require_relative in the specs to get the spec_helper.
This should be fine as messagebus doesn't appear to support
ruby 1.8.