Add PostgreSQL backend using LISTEN/NOTIFY
This supports PostgreSQL as a backend.  Originally, I attempted to
just provide a replacement Redis object that duck-typed and worked
on PostgreSQL, but due to differences between PostgreSQL and Redis
I was not able to get it to reliably pass the multi_process_spec.

This adds a lib/message_bus/postgres/reliable_pub_sub.rb file
with the PostgreSQL support.  To use the support, you currently
need to require that file manually, then pass a
:pub_sub_class=>MessageBus::Postgres::ReliablePubSub option in
your redis_config.  Connection options to PostgreSQL go in a
:pg_config subhash.

One significant difference between this PostgreSQL backend and the
Redis one is the dropping of channel-specific backlog ids. The
PostgreSQL backend only stores a single backlog id, not separate
global and channel backlog ids.  I don't think the lack of a
channel specific backlog id should be a problem. As long as the
channel backlog id is monotonically increasing, no messages
should be missed.

This does make some minor changes to the specs to make them support
both Redis and PostgreSQL.  Most of the changes are simple and
self explanatory.  I did add a ensure exit!(0) to the specs that
fork.  You can't call a normal exit in these specs, as the
garbage collector messes with the database connection sockets,
you need to call exit!(0) to skip that.  I added the ensure because
if there is an exception, it will be caught by minitest, but the
specs will keep running in both the child and parent processes.

Currently, the PostgreSQL backend does not automatically create the
table for you on first use, you need to do that manually by calling
MessageBus::Postgres::ReliablePubSub.setup!, which also clears the
table if it already exists.

This PostgreSQL backend reliably passes all specs for me, and works
for me in a simple application.  However, I didn't do any testing
with it at scale.