Remove redis runtime dependency from gemspec
Move it to a development dependency, and add a development
dependency on pg as well.

While here, set required_ruby_version in gemspec to 1.9.3.
MessageBus definitely does not support ruby 1.8.7 due at least to
use of ruby 1.9+ hash syntax.  I'm not sure if 1.9.2 is supported,
but since it's untested and probably unneeded, it shouldn't be an
issue.