Support passing channels to Rack middleware via env['message_bus.channels']
This changes MessageBus::Rack::Middleware to recognize the
'message_bus.channels' env entry as the channels to use for the
request, in preference to creating a new Rack::Request object
and using Rack::Request#POST.

This makes it easier for earlier middleware or request handler to
handle access controls for channels.  Currently, to do so would
require messing with rack.input, which is kind of a pain.

To support setting the client.seq value, that can also be specified
as the 'message_bus.seq' env entry.

There is no change to backwards compatibility if the
'message_bus.channels' env entry is not present.