Allow multi_process_spec to be run multiple times
This supports a MULTI_PROCESS_SPEC environment variable, specifying
how many times to run the spec.  As race conditions won't show up
on every run, you may need to run it multiple times, and this
provides an easy way to do so.  If one run fails, it skips the
remaining runs.

The other change is that after the responses have been collected,
the bus is unsubscribed normally, and the thread is joined.  This
makes it so threads aren't getting leaked.