Handle missing ActionDispatch::Flash in Railtie
The middleware config in Rails app operates on a MiddlewareStackProxy which more
or less acts as a recorder of the operations that is applied over default stack
when the app is initialized, hence a regular `rescue` won't work here.

There are two cases to cover:
1. Someone already "deleted" (meaning it will be deleted in the end but it is not just yet)
   ActionDispatch::Flash from the stack
2. ActionDispatch::Flash was never in the stack

 #1 happens with Rails4 apps that are configured to be API-only or if someone
doesn't want the middleware in their stack.
 #2 is a Rails5 API-only app behaviour.

To handle either case the initializer in `MessageBus::Rails::Railtie` is looking
for both cases and applies the MessageBus middleware appropriately.
The default behaviour can be easily changed via regular Rails config block.

```ruby
 # config/initializers/message_bus.rb
Rails.application.config do |config|
  # do anything you wish with config.middleware here
end
```