Makes forking spec simpler and less flaky
This test was introduced in https://github.com/SamSaffron/message_bus/commit/9c16e7ebaafaf2a3933a84fa1c517c0eba44b052 . Its purpose does not appear to be to assert that the forked process has a subscription (since no code is present to ensure that delivery ocurrs to both processes), nor does the order of publication or delivery appear to be important in this spec. All we're really interested in asserting here is that publications before a fork, from the forked process, and from the original process after the fork ocurrs, all get published and delivered. This change expresses that case much more simply.

Prior to this change, this test would fail on Redis approximately 20% of the time. With this change, it did not fail at all in 30 executions.