Tests backlog expiry actually working
The previous tests simply asserted that the backend-specific mechanism for expiring backlogs on Redis was set up, but not that it actually worked. The test could not, then, be exercised on other backends, whose implementations of backlog expiry do not have the same behaviour as those of the Redis backend.

These tests now assert the behaviour of the Redis backend, but do not invoke any private APIs specific to Redis, and can thus be applied to other backend implementations.