Only publish-trigger expiry for backends that require it
This allows the tests to assert that the preferred behaviour is for expiry to be automatic at the correct time, and that this ocurrs on compatible backends. Conversely, the tests assert that backends which require publication-triggered expiry do not auto-expire; if they gain such functionality in the future, the tests would have to be updated to reflect this, allowing us to work towards eliminating this divergence in behaviour from the tests.