on_middleware_error only for internal errors
Prior to this change, the `on_middleware_error` error handler was being invoked on all exceptions that hit this middleware, including those which come from lower-level middleware. This has the side-effect of potentially masking exceptions from the Rails application, and is surprising because `on_middleware_error` is assumed to handle only message_bus errors.