FIX: Protect Redis config from being manipulated
Redis gem is allowing to pass custom connector as an option. Later, code is removing that option and initialize custom connector:
```
options.delete(:connector).new(@options) # https://github.com/redis/redis-rb/blob/master/lib/redis/client.rb#L95
```

It works for MessageBus when Redis server is running. When it fails and MessageBus is trying to reconnect, because it is a passing reference to Redis, customer connector is not available anymore.

Therefore, I think it would be better to pass a copy of the original object.