FIX: Ensure non-long-polling requests are always spaced out (#324)
When long-polling, after a request returns some data, message-bus will immediately begin another long-polling request to maintain a consistent stream of messages.

When not long polling, we do not expect a continuous stream of messages, and the client aims to make one request every `backgroundCallbackInterval` (default 60s).

Before this commit, some long-polling logic was leaking into the non-long-polling requests. Under certain situations (e.g. a channel receiving very frequent messages), this could cause clients to start issuing non-long-polling requests every `minPollInterval` (100ms). This commit fixes the logic so that 'immediate retry' is only performed when in long-polling mode.

A number of tests are also introduced to demonstrate the logic and reduce the chance of future regressions.