FIX: Do not disable chunking on cancel/error
If the first chunk of a message takes more than 3s, we assume an intermediate proxy is buffering data and drop into dont-chunk mode. Previously, this 3s timer would continue even if a request was aborted before the first chunk arrived (e.g. channel subscriptions change in quick succession). This commit ensures the timer is cancelled when the request is aborted.