Don't error when trying to write a message to a closed client take 2 (#339)
This was first introduced in a2a46fde87f4c9c6cff806db5d84c2544befaa6d
but reverted because a new MessageBus release broke production and we
had to quickly revert before being able to isolate the commit. We have
since tested this commit in isolation in production and confirm that
MessageBus continues to work.

When a client is set to use chunked encoding by setting `use_chunked` to
true, trying to write a message to the client when the client has been
closed will raise an error. This commit changes it such that we don't
raise an error since the raised error doesn't do anything useful.